wip: restructure tracer around registry
This is to be split into separate OTP applications to make the
distinction clearer. For now the SDK registry is started when
opentelemetry application is started. The registry is responsible
for tracking the tracers, the sampler and span processors.

The registry also currently sets a ctx and span module to use. I
think this will change when context is split out.  I'd also like
to consider not having the span module be configurable in the
SDK but always use ot_span_ets and if the user needs a fancy form
of span handling they can replace the SDK if they can't get by
with using span processors.