metrics sdk first iteration
aggregation is instrument type specific until views are added to the spec.

releasing bound metrics does nothing and won't until the active instrument
table tracks if updates have occured within the last collection interval
for each active instrument and deletes them after 2 intervals of no
updates. At that point binding and releasing will need to be tracked
along with if there have been updates so it is known if an active instrument
is actually inactive and should be removed from the table or not.