make TracerProvider simply implement a gen_server
The API defines the functions a user calls to interact with the
TracerProvider but the handling of the gen_server calls is done
in another module. In this case the other module is in the SDK
application.

otel_tracer_server registers itself with the name
otel_tracer_provider so it acts as the default global provider.

If multiple SDK's attempt to register as the default providers
startup will fail, by design.