simple_processor: transition to idle after exporter timeout
Since the simple_processor is only exporting a single span and
does so based on a blocking call when a span ends there is no
need to do a `repeat_state` after the exporting time out fails.
The next export will be done if there is a span end message in
the processor's queue.