tracers: store startup created tracers the same way as user created
To simplify the understanding of tracer registration and fetching
this removes the separation between tracers registered on startup
for each application and those registered by a user or other code
at runtime.

Now there is a map to lookup the application name for a module
which is used to lookup the tracer, instead of before where the
map lookup actually got the tracer (meaning the tracer record was
duplicated and stored for every module!).

This means a user registering a tracer can override the tracer
registered at startup for an application name if they use the same
name, but I think this is sort of expected behavior so not an issue.

The change shouldn't effect much, if any, existing user code since
it almost always that tracers are fetched behind the scenes by
the macros.