limits on the number of links, events, attributes and their value size
Attribute limits can be set per event, link and span, as well as
general attribute limits that are used if no specific limits are
set for the individual type.

The number of attributes, links and events dropped do to being
over the limit is tracked and required changing each to be a
record that holds both the limits to apply and how many have
been dropped.