remove register_tracer. always use get_tracer
This fits the spec and other implementations of otel where the
term register is only used when talking about TracerProviders.
We don't currently expose a register for providers and I don't
know that we will. Providers are processes and I'd say them
starting and registering at atom name is what their "registering"
is.