fix use of traces_exporter app env var for setting the exporter
The atom `none` is now used instead of `undefined` for telling the
exporter to do nothing. This was done because `undefined` is used
in `otel_configuration` to signal that a variable is not set and
to not override it in the processor's config. That lead to the
`traces_exporter` value not being used if it was set to `none` or
`undefined` by the user.