use checkpoints of metrics for collection
In order to collect the current value of a metric, and reset it in
the case of a delta temporality, the reader process first must
checkpoint each metric. Checkpointing copies the current value to
a checkpoint field of the aggregation record and resets the value
and start time if needed. Then the checkpoint field is used to get
the metric value when collecting.

This is the easiest method to collect metrics without potentially
losing measurements during a reset that works across all types of
aggregations. There may be optimizations that can be done per
aggregation in the future.

The aggregation checkpoint function is called for each metric
individually, but this should likely change to have it so the
reader only has to call checkpoint for an aggregation and it
runs against all metrics in the table.