Fix OTLP Protocol allowed values
As the [specification describes][1], there are 3 possible values for ENV vars
that set the OTLP protocol used by OTLP exporter:
- grpc
- http/protobuf
- http/json

The SDK implementation leans on `list_to_existing_atom` function, so our
expected values are in fact `http_protobuf` instead of `http/protobuf`
for example.

This is confusing for ones following the generic documentation, and also
may become a problem in polyglot environments: let's say, I share the
same ENV setup for a Java and Erlang application.

The fix here account for the current atom values, and strings with the
same snake_case format for backwards compatibility. But also accept the
standard approach.

The fix includes `http/json` as possible transform, but the exporter
will still fails as that isn't supported by our impl.

[1]: https://github.com/open-telemetry/opentelemetry-specification/blob/82707fd9f7b1266f1246b02ff3e00bebdee6b538/specification/protocol/exporter.md#specify-protocol