move readers under metrics supervisor and require registered reader
This patch moves readers under a readers supervisor instead of being
started from the metrics server. Readers must be declared in config
and provided with a unique atom name that is used for registration
of the process. This name is also used to create named ets tables
to store the view aggregation mappings and the metrics.

Use of named ets tables and registered processes means the pid and
tids don't need to be tracked but a future update may find it
better to remove the need for names.