add logger handler which acts as LoggerProvider/Logger
This is a very simple version of what is needed to support the
logging spec and is in part duplicating code from the batch
span processor.

It has no overload protection and this will wait until we know
what is planned for log sampling in the spec as it will overlap.