remove SDK prep_stop noop tracer
Setting the default tracer to a noop in prep_stop wasn't only
wrong but also didn't really do anything since the tracer is
copied and not referenced, so existing tracers would remain
as is after setting the default to a noop.

But it is also wrong since we don't want to stop accepting traces
at any point and should try to create and export as many of
a users traces as we can before shutdown is complete.