use self() for grpc channel name
using ?MODULE limited us to a single grpcbox client channel. We
may want to not have a channel for each Signal (traces, meitrcs,
logs) if they are the same configuration, but for now this at
least allows each to export -- before only 1 could init their
exporter because it would conflict with the other.

Reuse of the channel is not as simple as using ?MODULE and if
it is already started use that channel because the endpoint may
be different. Not only that, the configuration to the endpoint
may be different, so a Channel name would have to be all the
distiguishing attributes of a Channel... Just easier to create
one for each Signal.