batch exporter: no longer rename table being exported
Until there is concurrent exporting there is no need for this.
Renaming can also mean a span is lost because it is to be
written to the table but that table then disappears. There is
still the chance this happens once the table is deleted before
being recreated, but it lessens it.