register Instruments with meter provider when created
when an Instrument is added it is matched against existing Views
to create a ViewAggregation for each match (and for each Reader).

ViewAggregations table is updated to be a bag to simplify the
updating of Instrument to ViewAggregation mapping. Now it is
just an insert for the newly matched Views if a View is added
and matches against existing Instruments.