inefficient use of matchspec for view matching
This patch uses ets:match_spec_run to compare a single View against
a single Instrument. This is to be the smallest possible change
when moving to use of a matchspec. The more efficient version will
come with running the matchspec against the Instruments table to
find matches. Though for new Instruments the inefficient version
will remain.