Store otel_resource attributes keys as atoms
Imported attribute keys (from opentelemetry_semantic_conventions) are
atoms, so otel_resource should support them.

As the resource is initialized only once and it is going to be copied
a lot, it's worth storing the provided keys as atoms to save memory.

Also, otel_resource_detector shouldn't know the type of the key, so add
an indirection level for the key check.