improve some typing thanks to eqwalizer
This has more dialyzer ignores added in order to remove stuff like
'$1' and '_' in type specs. I figure this is acceptable because
eqwalizer has a more fine grained ignore ability which will let us
check these parts of the code better even if the dialyzer warnings
are now ignored from those parts of the code.