metrics: support delta/cumulative conversions in sum aggregator
Cumulative aggregate is now also reset on each collection. A
previous checkpoint field is tracked in order to add the current
value to the previous in the case of needing the cumulative and
subtracting in the case you have a cumulative but need a delta
(example being an Observable which reports the cumulative value
in each callback and needed it reported as a delta).