create unique processor name in otel_tracer_server
Each processor needs a unique name to reference but always creating
it in start_link means everytime the process restarts it creates
a new atom and new persistent term key. This is a leak. This patch
moves the atom creation to otel_tracer_server so it is only done once.