update eqwalizer ignores
now in experimental we don't add all the matchspec types to the records
used in ets queries but instead use eqwalizer ignore comments above
anywhere the matchspec is used.

This allows for better typing of the actual values of the records
and eases the typing of the places that use those values, like
when a sum_aggregation record is extracted to a datapoint.