fix otel_span:start_opts typespec
fixes #708

In the API functions, all members of start_opts() are optional.

Before passing the options on the SDK all options are set to sensible
defaults. That makes is possible to has a sdk_start_opts() type where
all values are mandatory and optimize the usage of the opts slighty.