add otel_span:start_config typespec and use it in the SDK
fixes #708

In the API functions, all members of start_opts() are optional.

Before passing the options on the SDK all options are set to sensible
defaults. That makes is possible to have a dedicated start_config()
type where all values are mandatory and optimize the usage of the
settings slightly.