convert map of attributes to binary
bug in Erlang ETS allowed attributes as maps in select_replace
to appear to work when in fact they will match like a normal `=`
in Erlang, so any metric will be match by another with a subset
of its attributes. Converting to a binary means the attributes
must be exactly equal.