fix support for attribute lists of strings to not flatten them
No longer are attribute values that are lists of strings converted
to a single string. This means Erlang charlists will be sent in
OTLP as a list of integers and iolists as a list of strings. The
user must use characters_to_binary to convert to a binary string
if they wish to set the value to a string when they have a value
of those types now.