use new sdk configuration and move defaults
sdk_disabled flag from the old flat structure is converted to new
format from configuraiton SIG.

defaults are read where the configuration is read. this is so a
component can know the difference between a configuration key
being defined as empty and being undefined. before the default
was set for the key when configuration was read, meaning components
could not distinguish between if the configuration was empty,
undefined or set to the default by the user.