chore: update changelogs with PR numbers
also removes 2 chore commits that weren't marked as such from
being included using `--skip-commit`. going forward the
commit linter will keep unconventional commits out of the
history so these will be the only two that have to be ignored.

We couldn't use filter_unconventional = true to keep them out of
the history because 1 commit which needs to be included was not
a conventional commit.