chore: replace linting of individual commits with linting PR titles
This allows us to keep the same merge strategy (squash) as the
rest of the opentelemetry project while still verifying the commit
titles work with git-cliff.

It does not currently verify individual commits in a PR adhere to
the conventions. I'm torn on that. On the one hand I think they
should, but on the other it is unnecessary burden when everything
will be squashed. So I'm leaving it turned off for individual
commits.